+++ src/exekias/Metadata.cs
@@ -8,8 +8,7 @@ using System.Text.Json;
 partial class Program
 {
 
-    static async Task DoQuery(
-        InvocationContext ctx,
+    static async Task<int> DoQuery(
         FileInfo? cfgFile,
         string query,
         string orderBy,
@@ -21,8 +20,7 @@ partial class Program
         var cfg = LoadConfig(cfgFile, console);
         if (cfg == null)
         {
-            ctx.ExitCode = 1;
-            return;
+            return 1;
         }
         var exekiasStore = new ExekiasStore(Options.Create(new ExekiasStore.Options()
         {
@@ -48,7 +46,40 @@ partial class Program
                 Console.WriteLine(run.Run);
             }
         }
-            //.Select(exekiasObject => JObject.Parse(System.Text.Json.JsonSerializer.Serialize(exekiasObject)))
-        return;
+        //.Select(exekiasObject => JObject.Parse(System.Text.Json.JsonSerializer.Serialize(exekiasObject)))
+        return 0;
+    }
+
+    static async Task<int> DoShow(
+        FileInfo? cfgFile,
+        string runId,
+        IConsole console)
+    {
+        var cfg = LoadConfig(cfgFile, console);
+        if (cfg == null)
+        {
+            return 1;
+        }
+        var exekiasStore = new ExekiasStore(Options.Create(new ExekiasStore.Options()
+        {
+            ConnectionString = cfg.exekiasStoreConnectionString,
+            DatabaseName = cfg.exekiasStoreDatabaseName,
+            ContainerName = cfg.exekiasStoreContainerName
+        }), LoggerFactory.Create(builder =>
+        {
+            builder
+                .AddFilter("Exekias.CosmosDB", LogLevel.Error)
+                .AddConsole();
+        }).CreateLogger<ExekiasStore>());
+        var result = await exekiasStore.GetMetaObject(runId);
+        if (null == result)
+        {
+            Error(console, $"Run '{runId}' cannot be found.");
+        }
+        else
+        {
+            Console.WriteLine(JsonSerializer.Serialize(result, new JsonSerializerOptions() { WriteIndented = true }));
+        }
+        return 0;
     }
 }