+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+- `exekias query ...` command is now `exekias runs query ...`
+
+### Added
+- `exekias runs show <run>` command to print run metadata.
+
+
 ## [2.0.5] - 2023-09-26
 ## [2.0.4] - 2023-09-26